Add email-notifications-plugin
