+++ config/warble.rb
@@ -20,27 +20,3 @@ Warbler::Config.new do |config|
 
   config.jar_name = 'killbill-admin-ui-standalone'
 end
-
-class Warbler::Jar
-  def add_init_file(config)
-    return unless config.init_contents
-
-    contents = +''
-    config.init_contents.each do |file|
-      if file.respond_to?(:read)
-        contents << file.read
-      elsif File.extname(file) == '.erb'
-        contents << expand_erb(file, config).read
-      else
-        contents << File.read(file)
-      end
-    end
-    @files[config.init_filename] = StringIO.new(contents)
-    return unless @files[config.init_filename]
-
-    contents = @files[config.init_filename].read
-    add_before = "require 'bundler/shared_helpers'"
-    line_to_add = "Gem.paths = ENV\n"
-    @files[config.init_filename] = StringIO.new(contents.gsub(add_before, "#{line_to_add}\\0"))
-  end
-end
